Move simple_post() closer to where it's used
