+++ corehq/motech/requests.py
@@ -6,6 +6,7 @@ from django.conf import settings
 import attr
 import requests
 from requests.auth import HTTPBasicAuth, HTTPDigestAuth
+from requests.structures import CaseInsensitiveDict
 
 from dimagi.utils.logging import notify_exception
 
@@ -215,3 +216,31 @@ def parse_request_exception(err):
     else:
         err_response = str(err)
     return err_request, err_response
+
+
+def simple_post(domain, url, data, *, headers, auth, verify,
+                timeout=60, notify_addresses=None, payload_id=None):
+    """
+    POST with a cleaner API, and return the actual HTTPResponse object, so
+    that error codes can be interpreted.
+    """
+    default_headers = CaseInsensitiveDict({
+        "content-type": "text/xml",
+        "content-length": str(len(data)),
+    })
+    default_headers.update(headers)
+    kwargs = {
+        "headers": default_headers,
+        "timeout": timeout,
+    }
+    requests = Requests(
+        domain,
+        base_url='',
+        username=None,
+        password=None,
+        verify=verify,
+        notify_addresses=notify_addresses,
+        payload_id=payload_id,
+    )
+    # Use ``send_request()`` instead of ``post()`` to pass ``auth``.
+    return requests.send_request('POST', url, data=data, auth=auth, **kwargs)
